added functioning ground detection script and make jumping better
added script to physical player gameobject that detects if the player is on the ground. can be expanded to see what type of gound the player is on. also make the player not able to jump if already in the air. just need to check later if the jump is consistent.

next thing im gonna work on is the camera and the player control of it